Restructuring backend code base
