+++ backend/vehicles/src/modules/power-unit-types/dto/request/create-power-unit-type.dto.ts
