+++ backend/vehicles/src/modules/power-unit-types/dto/request/update-power-unit-type.dto.ts
