+++ backend/vehicles/src/modules/power-unit-types/dto/response/read-power-unit-type.dto.ts
