+++ backend/vehicles/src/modules/power-unit-types/power-unit-types.service.spec.ts
