+++ backend/vehicles/src/modules/power-unit-types/profiles/power-unit-type.profile.ts
