+++ backend/vehicles/src/modules/power-units/dto/response/read-power-unit.dto.ts
