+++ backend/vehicles/src/modules/power-units/power-units.controller.ts
@@ -20,8 +20,8 @@ import {
 } from '@nestjs/swagger';
 
 import { ReadPowerUnitDto } from './dto/response/read-power-unit.dto';
-import { ExceptionDto } from '../common/dto/exception.dto';
-import { DataNotFoundException } from '../common/exception/data-not-found.exception';
+import { ExceptionDto } from '../../common/dto/exception.dto';
+import { DataNotFoundException } from '../../common/exception/data-not-found.exception';
 
 @ApiTags('Power Units')
 @ApiNotFoundResponse({
