+++ backend/vehicles/src/modules/power-units/power-units.service.ts
