+++ backend/vehicles/src/modules/power-units/profiles/power-unit.profile.ts
