+++ backend/vehicles/src/modules/trailer-types/dto/request/create-trailer-type.dto.ts
