+++ backend/vehicles/src/modules/trailer-types/dto/request/update-trailer-type.dto.ts
