+++ backend/vehicles/src/modules/trailer-types/dto/response/read-trailer-type.dto.ts
