+++ backend/vehicles/src/modules/trailer-types/entities/trailer-type.entity.ts
@@ -1,7 +1,8 @@
 import { AutoMap } from '@automapper/classes';
 import { ApiProperty } from '@nestjs/swagger';
 import { Entity, Column, OneToMany, PrimaryColumn } from 'typeorm';
-import { Base } from '../../common/entities/base.entity';
+import { Base } from '../../../common/entities/base.entity';
+
 import { Trailer } from '../../trailers/entities/trailer.entity';
 
 @Entity({ name: 'ORBC_VT_TRAILER_TYPE' })
