+++ backend/vehicles/src/modules/trailer-types/profiles/trailer-type.profile.ts
