+++ backend/vehicles/src/modules/trailer-types/trailer-types.module.ts
