+++ backend/vehicles/src/modules/trailer-types/trailer-types.service.spec.ts
