+++ backend/vehicles/src/modules/trailer-types/trailer-types.service.ts
