+++ backend/vehicles/src/modules/trailers/dto/request/create-trailer.dto.ts
