+++ backend/vehicles/src/modules/trailers/dto/request/update-trailer.dto.ts
