+++ backend/vehicles/src/modules/trailers/dto/response/read-trailer.dto.ts
