+++ backend/vehicles/src/modules/trailers/trailers.controller.ts
@@ -20,8 +20,8 @@ import {
 } from '@nestjs/swagger';
 
 import { ReadTrailerDto } from './dto/response/read-trailer.dto';
-import { ExceptionDto } from '../common/dto/exception.dto';
-import { DataNotFoundException } from '../common/exception/data-not-found.exception';
+import { ExceptionDto } from '../../common/dto/exception.dto';
+import { DataNotFoundException } from '../../common/exception/data-not-found.exception';
 
 @ApiTags('Trailers')
 @ApiNotFoundResponse({
