+++ backend/vehicles/src/modules/trailers/trailers.module.ts
