+++ backend/vehicles/src/modules/trailers/trailers.service.ts
