+++ backend/vehicles/test/trailer-types.e2e-spec.ts
@@ -8,11 +8,11 @@ import { AutomapperModule } from '@automapper/nestjs';
 
 import { createMock } from '@golevelup/ts-jest';
 import { Repository } from 'typeorm';
-import { CreateTrailerTypeDto } from '../src/trailer-types/dto/request/create-trailer-type.dto';
-import { UpdateTrailerTypeDto } from '../src/trailer-types/dto/request/update-trailer-type.dto';
-import { ReadTrailerTypeDto } from '../src/trailer-types/dto/response/read-trailer-type.dto';
-import { TrailerType } from '../src/trailer-types/entities/trailer-type.entity';
-import { TrailerTypesModule } from '../src/trailer-types/trailer-types.module';
+import { CreateTrailerTypeDto } from '../src/modules/trailer-types/dto/request/create-trailer-type.dto';
+import { UpdateTrailerTypeDto } from '../src/modules/trailer-types/dto/request/update-trailer-type.dto';
+import { ReadTrailerTypeDto } from '../src/modules/trailer-types/dto/response/read-trailer-type.dto';
+import { TrailerType } from '../src/modules/trailer-types/entities/trailer-type.entity';
+import { TrailerTypesModule } from '../src/modules/trailer-types/trailer-types.module';
 
 const typeCode = 'BOOSTER';
 const type = 'Boosters';
